requirements.txt
aiohttp==3.8.5
customtkinter==5.2.2
pandas==1.5.2
pillow==10.2.0
beautifulsoup4==4.12.2
aiohttp-retry==2.8.3